Move abstracts classes to Abstractions namespace
