+++ src/Sumodesk.Domain/Abstractions/EntityRoot.cs
@@ -1,6 +1,6 @@
 ﻿using Sumodesk.Domain.Interfaces;
 
-namespace Sumodesk.Domain.Entities;
+namespace Sumodesk.Domain.Abstractions;
 
 public abstract class EntityRoot : Entity, IEntityRoot
 {
